refactor: 更新数据模型和数据库访问层，调整主键类型为 Integer，优化表结构

+++ avalon-application/avalon-admin-server/src/main/java/io/github/lishangbu/avalon/admin/service/dataset/BerryFirmnessService.java
@@ -49,7 +49,7 @@ public interface BerryFirmnessService {
    *
    * @param id 树果坚硬度ID
    */
-  void removeById(Long id);
+  void removeById(Integer id);
 
   /**
    * 根据条件查询树果坚硬度列表
